Add racey functions for using your own RNG seed
The race functions were not available on no_std targets. This commit
adds "with_seed" variants that take a u64 seed that is passed to
fastrand. It can be used on no_std targets to provide random racey
futures and streams.

A "race" feature is added that enables it. This way fastrand doesn't
become a dependency for crates that don't need it, like async-io.

Signed-off-by: John Nunley <dev@notgull.net>